formatted and documented and moved tests
