+++ src/common/data.js
@@ -1,8 +1,4 @@
-function buildGetData({
-  nilID,
-  encodeID,
-  deps,
-}) {
+exports.buildGetData = ({ nilID, encodeID, deps }) => {
   return function getData(auth, ids, refresh, resume) {
     if (encodeID(ids) === nilID || ids.length === 3 && +ids[2] === 0) {
       resume(null, {});
@@ -11,5 +7,4 @@ function buildGetData({
       deps.compileID(auth, id, {refresh: refresh}, resume);
     }
   };
-}
-exports.buildGetData = buildGetData;
+};
