+++ src/static/storers/gcs.test.js
@@ -1,5 +1,7 @@
 describe('storer/gcs', () => {
   it('should have tests', () => {
-    
+    // Arrange
+    // Act
+    // Assert
   });
 });
