+++ src/static/storers/throwing.test.js
